[#31] fix: style
