+++ gatsby-node.js
@@ -12,7 +12,7 @@ exports.createPages = async ({
 
   // Define a template for blog post
   const blogPost = path.resolve(
-    `./src/components/templates/blog-post.tsx`,
+    `./src/templates/blog-post.tsx`,
   );
 
   // Get all markdown blog posts sorted by date
