+++ src/components/templates/layout.tsx
@@ -140,140 +140,6 @@ const GlobalWrapper = styled.div`
   .by-section-foot {
     margin-top: 24px;
   }
-
-  .by-btn {
-    display: inline-flex;
-    position: relative;
-    align-items: center;
-    justify-content: center;
-    flex-direction: row;
-    cursor: pointer;
-    line-height: 1;
-    padding: 10px 16px;
-    font-weight: 400;
-    white-space: nowrap;
-    border: 0;
-    border-radius: 8px;
-    color: ${(props) => props.theme.colors.background};
-    border: 1px solid ${(props) => props.theme.colors.base};
-    background: ${(props) => props.theme.colors.base};
-    transition: ${(props) =>
-        props.theme.colorModeTransition},
-      box-shadow 0.25s var(--ease-in-out-quad),
-      transform 0.25s var(--ease-in-out-quad);
-
-    &:hover {
-      color: ${(props) => props.theme.colors.base};
-      border: 1px solid
-        ${(props) => props.theme.colors.base};
-      background: transparent;
-    }
-
-    svg {
-      fill: ${(props) => props.theme.colors.light};
-      margin-left: 4px;
-    }
-
-    &.by-btn-outline {
-      color: ${(props) => props.theme.colors.light};
-      border: 1px solid
-        ${(props) => props.theme.colors.light};
-      background: transparent;
-
-      &:hover {
-        color: ${(props) => props.theme.colors.base};
-        border: 1px solid
-          ${(props) => props.theme.colors.base};
-
-        svg {
-          fill: ${(props) => props.theme.colors.base};
-        }
-      }
-    }
-  }
-
-  .tooltip-container {
-    position: relative;
-
-    .tooltip-text {
-      position: absolute;
-      display: inline-block;
-      font-size: 0.75rem;
-      padding: 4px 8px;
-      border-radius: 4px;
-      color: ${(props) => props.theme.colors.background};
-      background: ${(props) => props.theme.colors.base};
-      white-space: nowrap;
-      opacity: 0;
-      visibility: hidden;
-      transition: opacity 0.25s var(--ease-in-out-quad),
-        visibility 0.25s var(--ease-in-out-quad);
-
-      &::after {
-        content: ' ';
-        position: absolute;
-        border-width: 5px;
-        border-style: solid;
-        border-color: ${(props) => props.theme.colors.base}
-          transparent transparent transparent;
-      }
-    }
-
-    .tooltip-top {
-      bottom: calc(100% + 8px);
-      left: 50%;
-      transform: translateX(-50%);
-
-      &::after {
-        top: 100%;
-        left: 50%;
-        transform: translateX(-5px);
-      }
-    }
-
-    .tooltip-bottom {
-      top: calc(100% + 8px);
-      left: 50%;
-      transform: translateX(-50%);
-
-      &::after {
-        bottom: 100%;
-        left: 50%;
-        transform: translateX(-5px) rotate(180deg);
-      }
-    }
-
-    .tooltip-right {
-      left: calc(100% + 8px);
-      top: 50%;
-      transform: translateY(-50%);
-
-      &::after {
-        right: 100%;
-        top: 50%;
-        transform: translateY(-5px) rotate(90deg);
-      }
-    }
-
-    .tooltip-left {
-      right: calc(100% + 8px);
-      top: 50%;
-      transform: translateY(-50%);
-
-      &::after {
-        left: 100%;
-        top: 50%;
-        transform: translateY(-5px) rotate(-90deg);
-      }
-    }
-
-    &:hover {
-      .tooltip-text {
-        opacity: 1;
-        visibility: visible;
-      }
-    }
-  }
 `;
 
 const Main = styled.main`
