+++ src/pages/contact.tsx
@@ -15,138 +15,7 @@ import styled from '../components/atoms/styled';
 import Layout from '../components/templates/layout';
 import SEO from '../components/organisms/seo';
 import PageHero from '../components/organisms/hero';
-
-const FormWrapper = styled.div`
-  form {
-    display: grid;
-    grid-template-columns: 200px 1fr;
-    grid-gap: 16px;
-
-    label {
-      padding: 8px 0px 0px 0px;
-      margin-bottom: -8px;
-
-      &.option {
-        color: ${(props) => props.theme.colors.light};
-      }
-    }
-
-    label,
-    .form-field,
-    button {
-      grid-column: 1 / 3;
-    }
-
-    .error-msg {
-      color: red;
-      padding: 8px 16px 0px;
-    }
-
-    @media screen and (min-width: ${(props) =>
-        props.theme.responsive.medium}) {
-      label {
-        grid-column: 1 / 2;
-        padding: 8px 0px;
-      }
-
-      .form-field,
-      button {
-        grid-column: 2 / 3;
-      }
-    }
-
-    input,
-    textarea,
-    button,
-    select {
-      display: block;
-      width: 100%;
-      appearance: none;
-      border: 1px solid
-        ${(props) => props.theme.colors.border};
-      border-radius: 8px;
-      padding: 8px 16px;
-      color: ${(props) => props.theme.colors.base};
-      background: ${(props) =>
-        props.theme.colors.backgroundAccent};
-      transition: ${(props) =>
-        props.theme.colorModeTransition};
-
-      &:focus {
-        outline: none;
-        border: 1px solid
-          ${(props) => props.theme.colors.base};
-      }
-    }
-
-    .select-wrapper {
-      position: relative;
-
-      &::after {
-        content: '';
-        display: block;
-        position: absolute;
-        top: 50%;
-        right: 16px;
-        width: 6px;
-        height: 6px;
-        border-bottom: solid 2px
-          ${(props) => props.theme.colors.light};
-        border-right: solid 2px
-          ${(props) => props.theme.colors.light};
-        transform: translateY(-50%) rotate(45deg);
-      }
-    }
-
-    .by-btn {
-      svg.send-icon,
-      svg.load-icon,
-      svg.success-icon,
-      svg.fail-icon {
-        width: 16px;
-        height: 16px;
-      }
-
-      svg.send-icon,
-      svg.load-icon {
-        fill: ${(props) => props.theme.colors.background};
-      }
-
-      svg.load-icon {
-        animation: 2s linear infinite rotation;
-      }
-
-      @keyframes rotation {
-        0% {
-          transform: rotate(0);
-        }
-        100% {
-          transform: rotate(-360deg);
-        }
-      }
-
-      svg.success-icon {
-        fill: green;
-      }
-
-      svg.fail-icon {
-        fill: red;
-      }
-
-      &:hover {
-        svg.send-icon,
-        svg.load-icon {
-          fill: ${(props) => props.theme.colors.base};
-        }
-      }
-    }
-  }
-`;
-
-const IconWrapper = styled.div`
-  line-height: 0;
-  margin-right: 4px;
-`;
+import Button from '../components/atoms/button';
 
 const CONTACT_CATEGORIES = [
   'Web制作を依頼したい',
@@ -302,77 +171,65 @@ const ContactIndex: React.FC<
           <ErrorMessage name="message" />
         </div>
       </div>
-      <button
-        className="by-btn by-btn-primary"
+      <Button
+        startIcon={
+          formState === 'loading' ? (
+            <svg
+              className="load-icon"
+              xmlns="http://www.w3.org/2000/svg"
+              height="14"
+              viewBox="0 0 24 24"
+              width="14"
+            >
+              <path d="M0 0h24v24H0z" fill="none" />
+              <path d="M12 4V1L8 5l4 4V6c3.31 0 6 2.69 6 6 0 1.01-.25 1.97-.7 2.8l1.46 1.46C19.54 15.03 20 13.57 20 12c0-4.42-3.58-8-8-8zm0 14c-3.31 0-6-2.69-6-6 0-1.01.25-1.97.7-2.8L5.24 7.74C4.46 8.97 4 10.43 4 12c0 4.42 3.58 8 8 8v3l4-4-4-4v3z" />
+            </svg>
+          ) : formState === 'success' ? (
+            <svg
+              className="success-icon"
+              xmlns="http://www.w3.org/2000/svg"
+              height="14"
+              viewBox="0 0 24 24"
+              width="14"
+            >
+              <path d="M0 0h24v24H0z" fill="none" />
+              <path d="M9 16.17L4.83 12l-1.42 1.41L9 19 21 7l-1.41-1.41z" />
+            </svg>
+          ) : formState === 'fail' ? (
+            <svg
+              className="fail-icon"
+              xmlns="http://www.w3.org/2000/svg"
+              height="14"
+              viewBox="0 0 24 24"
+              width="14"
+            >
+              <path d="M0 0h24v24H0z" fill="none" />
+              <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm1 15h-2v-6h2v6zm0-8h-2V7h2v2z" />
+            </svg>
+          ) : (
+            <svg
+              className="send-icon"
+              xmlns="http://www.w3.org/2000/svg"
+              height="14"
+              viewBox="0 0 24 24"
+              width="14"
+            >
+              <path d="M0 0h24v24H0z" fill="none" />
+              <path d="M2.01 21L23 12 2.01 3 2 10l15 2-15 2z" />
+            </svg>
+          )
+        }
         type="submit"
         disabled={isSubmitting}
       >
-        {formState === 'loading' ? (
-          <>
-            <IconWrapper>
-              <svg
-                className="load-icon"
-                xmlns="http://www.w3.org/2000/svg"
-                height="24"
-                viewBox="0 0 24 24"
-                width="24"
-              >
-                <path d="M0 0h24v24H0z" fill="none" />
-                <path d="M12 4V1L8 5l4 4V6c3.31 0 6 2.69 6 6 0 1.01-.25 1.97-.7 2.8l1.46 1.46C19.54 15.03 20 13.57 20 12c0-4.42-3.58-8-8-8zm0 14c-3.31 0-6-2.69-6-6 0-1.01.25-1.97.7-2.8L5.24 7.74C4.46 8.97 4 10.43 4 12c0 4.42 3.58 8 8 8v3l4-4-4-4v3z" />
-              </svg>
-            </IconWrapper>
-            {`送信しています`}
-          </>
-        ) : formState === 'success' ? (
-          <>
-            <IconWrapper>
-              <svg
-                className="success-icon"
-                xmlns="http://www.w3.org/2000/svg"
-                height="24"
-                viewBox="0 0 24 24"
-                width="24"
-              >
-                <path d="M0 0h24v24H0z" fill="none" />
-                <path d="M9 16.17L4.83 12l-1.42 1.41L9 19 21 7l-1.41-1.41z" />
-              </svg>
-            </IconWrapper>
-            {`送信に成功しました`}
-          </>
-        ) : formState === 'fail' ? (
-          <>
-            <IconWrapper>
-              <svg
-                className="fail-icon"
-                xmlns="http://www.w3.org/2000/svg"
-                height="24"
-                viewBox="0 0 24 24"
-                width="24"
-              >
-                <path d="M0 0h24v24H0z" fill="none" />
-                <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm1 15h-2v-6h2v6zm0-8h-2V7h2v2z" />
-              </svg>
-            </IconWrapper>
-            {`送信に失敗しました`}
-          </>
-        ) : (
-          <>
-            <IconWrapper>
-              <svg
-                className="send-icon"
-                xmlns="http://www.w3.org/2000/svg"
-                height="24"
-                viewBox="0 0 24 24"
-                width="24"
-              >
-                <path d="M0 0h24v24H0z" fill="none" />
-                <path d="M2.01 21L23 12 2.01 3 2 10l15 2-15 2z" />
-              </svg>
-            </IconWrapper>
-            {`送信`}
-          </>
-        )}
-      </button>
+        {formState === 'loading'
+          ? '送信しています'
+          : formState === 'success'
+          ? '送信に成功しました'
+          : formState === 'fail'
+          ? '送信に成功しました'
+          : '送信'}
+      </Button>
     </Form>
   );
 
@@ -417,6 +274,123 @@ export const pageQuery = graphql`
         title
         social {
           twitter
+        }
+      }
+    }
+  }
+`;
+
+const FormWrapper = styled.div`
+  form {
+    display: grid;
+    grid-template-columns: 200px 1fr;
+    grid-gap: 16px;
+
+    label {
+      padding: 8px 0px 0px 0px;
+      margin-bottom: -8px;
+
+      &.option {
+        color: ${(props) => props.theme.colors.light};
+      }
+    }
+
+    label,
+    .form-field,
+    button {
+      grid-column: 1 / 3;
+    }
+
+    .error-msg {
+      color: red;
+      padding: 8px 16px 0px;
+    }
+
+    @media screen and (min-width: ${(props) =>
+        props.theme.responsive.medium}) {
+      label {
+        grid-column: 1 / 2;
+        padding: 8px 0px;
+      }
+
+      .form-field,
+      button {
+        grid-column: 2 / 3;
+      }
+    }
+
+    input,
+    textarea,
+    button,
+    select {
+      display: block;
+      width: 100%;
+      appearance: none;
+      border: 1px solid
+        ${(props) => props.theme.colors.border};
+      border-radius: 8px;
+      padding: 8px 16px;
+      color: ${(props) => props.theme.colors.base};
+      background: ${(props) =>
+        props.theme.colors.backgroundAccent};
+      transition: ${(props) =>
+        props.theme.colorModeTransition};
+
+      &:focus {
+        outline: none;
+        border: 1px solid
+          ${(props) => props.theme.colors.base};
+      }
+    }
+
+    .select-wrapper {
+      position: relative;
+
+      &::after {
+        content: '';
+        display: block;
+        position: absolute;
+        top: 50%;
+        right: 16px;
+        width: 6px;
+        height: 6px;
+        border-bottom: solid 2px
+          ${(props) => props.theme.colors.light};
+        border-right: solid 2px
+          ${(props) => props.theme.colors.light};
+        transform: translateY(-50%) rotate(45deg);
+      }
+    }
+
+    button {
+      svg.load-icon {
+        animation: 2s linear infinite rotation;
+      }
+
+      @keyframes rotation {
+        0% {
+          transform: rotate(0);
+        }
+        100% {
+          transform: rotate(-360deg);
+        }
+      }
+
+      svg.success-icon {
+        fill: green;
+      }
+
+      svg.fail-icon {
+        fill: red;
+      }
+
+      &:hover {
+        svg.success-icon {
+          fill: green;
+        }
+
+        svg.fail-icon {
+          fill: red;
         }
       }
     }
